# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ArogyaGenie Healthcare Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /lab-reports:
    get:
      operationId: listLabReports
      tags: [lab-reports]
      summary: List lab reports for current patient
      responses:
        "200":
          description: List of lab reports
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LabReport"
    post:
      operationId: createLabReport
      tags: [lab-reports]
      summary: Upload a lab report
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LabReportInput"
      responses:
        "201":
          description: Created lab report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LabReport"
  /lab-reports/{id}:
    get:
      operationId: getLabReport
      tags: [lab-reports]
      summary: Get lab report by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Lab report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LabReport"
    patch:
      operationId: updateLabReport
      tags: [lab-reports]
      summary: Update lab report (add AI summary, results)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LabReportUpdate"
      responses:
        "200":
          description: Updated lab report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LabReport"
  /lab-reports/{id}/analyze:
    post:
      operationId: analyzeLabReport
      tags: [lab-reports]
      summary: Trigger AI analysis for a lab report
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Analyzed lab report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LabReport"

  # ─── MEDICAL KNOWLEDGE (RAG) ───────────────────────────────────────────────
...
  /doctors/patients/{patientId}/ai-summary:
    get:
      operationId: getDoctorPatientAiSummary
      tags: [health-intelligence]
      summary: Get clinical AI briefing for authorized doctor consultation
      parameters:
        - name: patientId
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Doctor patient briefing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DoctorBriefing"
...
  /medicine-reminders:
    get:
      operationId: listMedicineReminders
      tags: [medicine-reminders]
      summary: List medicine reminders for current patient
      responses:
        "200":
          description: List of reminders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MedicineReminder"
    post:
      operationId: createMedicineReminder
      tags: [medicine-reminders]
      summary: Create a medicine reminder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MedicineReminderInput"
      responses:
        "201":
          description: Created reminder
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MedicineReminder"
  /medicine-reminders/{id}:
    patch:
      operationId: updateMedicineReminder
      tags: [medicine-reminders]
      summary: Update a medicine reminder
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MedicineReminderUpdate"
      responses:
        "200":
          description: Updated reminder
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MedicineReminder"
    delete:
      operationId: deleteMedicineReminder
      tags: [medicine-reminders]
      summary: Delete a medicine reminder
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Deleted

  # ─── TIMELINE ───────────────────────────────────────────────────────────────
...
  /pharmacies/me/profile:
    get:
      operationId: getPharmacyProfile
      tags: [pharmacies]
      summary: Get current pharmacy profile
      responses:
        "200":
          description: Pharmacy profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Pharmacy"
    put:
      operationId: updatePharmacyProfile
      tags: [pharmacies]
      summary: Update pharmacy profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PharmacyUpdate"
      responses:
        "200":
          description: Updated pharmacy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Pharmacy"
...
  /pharmacies/me/prescriptions:
    get:
      operationId: listPharmacyPrescriptions
      tags: [pharmacies]
      summary: List prescriptions for current pharmacy
      parameters:
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of prescriptions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Prescription"

  # ─── ADMIN ──────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Prescription:
      type: object
      properties:
        id:
          type: number
        patientId:
          type: number
        doctorId:
          type: number
        appointmentId:
          type: ["number", "null"]
        patientName:
          type: ["string", "null"]
        doctorName:
          type: ["string", "null"]
        medicines:
          type: string
        diagnosis:
          type: ["string", "null"]
        instructions:
          type: ["string", "null"]
        fileUrl:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, dispensed, expired]
        prescribedDate:
          type: string
        createdAt:
          type: string
      required:
        - id
        - patientId
        - doctorId
        - medicines
        - status
        - prescribedDate
        - createdAt
...
    LabReport:
      type: object
      properties:
        id:
          type: number
        patientId:
          type: number
        diagnosticCenterId:
          type: ["number", "null"]
        testName:
          type: string
        testDate:
          type: string
        fileUrl:
          type: ["string", "null"]
        aiSummary:
          type: ["string", "null"]
        results:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, completed, reviewed]
        createdAt:
          type: string
      required:
        - id
        - patientId
        - testName
        - testDate
        - status
        - createdAt
    LabReportInput:
      type: object
      required:
        - testName
        - testDate
      properties:
        testName:
          type: string
        testDate:
          type: string
        fileUrl:
          type: string
        results:
          type: string
    LabReportUpdate:
      type: object
      properties:
        aiSummary:
          type: string
        results:
          type: string
        status:
          type: string
          enum: [pending, completed, reviewed]
        fileUrl:
          type: string

    # ─── DIAGNOSTIC CENTER ─────────────────────────────────────────────────────
...
    MedicineReminder:
      type: object
      properties:
        id:
          type: number
        patientId:
          type: number
        medicineName:
          type: string
        dosage:
          type: string
        frequency:
          type: string
          enum: [once_daily, twice_daily, thrice_daily, every_8_hours, weekly, as_needed]
        times:
          type: string
        startDate:
          type: string
        endDate:
          type: ["string", "null"]
        instructions:
          type: ["string", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - patientId
        - medicineName
        - dosage
        - frequency
        - times
        - startDate
        - isActive
        - createdAt
    MedicineReminderInput:
      type: object
      required:
        - medicineName
        - dosage
        - frequency
        - times
        - startDate
      properties:
        medicineName:
          type: string
        dosage:
          type: string
        frequency:
          type: string
          enum: [once_daily, twice_daily, thrice_daily, every_8_hours, weekly, as_needed]
        times:
          type: string
        startDate:
          type: string
        endDate:
          type: string
        instructions:
          type: string
    MedicineReminderUpdate:
      type: object
      properties:
        medicineName:
          type: string
        dosage:
          type: string
        frequency:
          type: string
          enum: [once_daily, twice_daily, thrice_daily, every_8_hours, weekly, as_needed]
        times:
          type: string
        endDate:
          type: string
        instructions:
          type: string
        isActive:
          type: boolean

    # ─── TIMELINE ──────────────────────────────────────────────────────────────
...
    Pharmacy:
      type: object
      properties:
        id:
          type: number
        userId:
          type: number
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        licenseNumber:
          type: ["string", "null"]
        openingHours:
          type: ["string", "null"]
        status:
          type: string
      required:
        - id
        - userId
        - name
        - email
        - status
    PharmacyUpdate:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        address:
          type: string
        city:
          type: string
        licenseNumber:
          type: string
        openingHours:
          type: string
...
    DoctorBriefing:
      type: object
      properties:
        patientOverview:
          type: string
        keyDiagnoses:
          type: array
          items:
            type: string
        activeMedications:
          type: array
          items:
            type: string
        recentLabHighlights:
          type: array
          items:
            type: string
        suggestedFollowUpTopics:
          type: array
          items:
            type: string
        disclaimer:
          type: string
      required:
        - patientOverview
        - keyDiagnoses
        - activeMedications
        - recentLabHighlights
        - suggestedFollowUpTopics
        - disclaimer
...